requirements.txt
django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9


